chore: move functions to separate files
