+++ cmd/internal/kube/daemonset/daemonset.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"runtime/debug"
 	"sync"
 	"time"
 
-	"github.com/akitasoftware/go-utils/optionals"
 	"github.com/pkg/errors"
-	"github.com/postmanlabs/postman-insights-agent/apidump"
 	"github.com/postmanlabs/postman-insights-agent/apispec"
 	"github.com/postmanlabs/postman-insights-agent/integrations/cri_apis"
 	"github.com/postmanlabs/postman-insights-agent/integrations/kube_apis"
@@ -150,7 +147,7 @@ func (d *Daemonset) getPodArgsFromMap(podUID types.UID) (*PodArgs, error) {
 }
 
 // addPodArgsToMap adds the podArgs to the map with the podUID as the key
-// This funciton ensures that the pod is not already loaded in the map
+// This function ensures that the pod is not already loaded in the map
 func (d *Daemonset) addPodArgsToMap(podUID types.UID, args *PodArgs, startingState PodTrafficMonitorState) {
 	value, loaded := d.PodArgsByNameMap.LoadOrStore(podUID, args)
 	argsFromMap := value.(*PodArgs)
@@ -162,16 +159,6 @@ func (d *Daemonset) addPodArgsToMap(podUID types.UID, args *PodArgs, startingSta
 		}
 	} else {
 		printer.Errorf("Pod is already loaded in the map and is in state %d", argsFromMap.PodTrafficMonitorState)
-	}
-}
-
-func (d *Daemonset) sendTelemetry() {
-	ctx, cancel := context.WithTimeout(context.Background(), apiContextTimeout)
-	defer cancel()
-
-	err := d.FrontClient.PostDaemonsetAgentTelemetry(ctx, d.ClusterName)
-	if err != nil {
-		printer.Errorf("Failed to send telemetry: %v\n", err)
 	}
 }
 
@@ -254,44 +241,6 @@ func (d *Daemonset) KubernetesEventsWorker(done <-chan struct{}) {
 	}
 }
 
-func (d *Daemonset) checkPodsHealth() {
-	var podUIDs []types.UID
-	d.PodArgsByNameMap.Range(func(k, _ interface{}) bool {
-		podUIDs = append(podUIDs, k.(types.UID))
-		return true
-	})
-
-	podStatuses, err := d.KubeClient.GetPodsStatusByUIDs(podUIDs)
-	if err != nil {
-		printer.Errorf("Failed to get pods status: %v", err)
-		return
-	}
-
-	for podUID, podStatus := range podStatuses {
-		if podStatus == coreV1.PodSucceeded || podStatus == coreV1.PodFailed {
-			printer.Infof("Pod %s has stopped running", podStatus)
-
-			podArgs, err := d.getPodArgsFromMap(podUID)
-			if err != nil {
-				printer.Errorf("Failed to get podArgs for podUID %s: %v", podUID, err)
-				continue
-			}
-
-			err = podArgs.changePodTrafficMonitorState(PodTerminated, TrafficMonitoringStarted)
-			if err != nil {
-				printer.Errorf("Failed to change pod state, pod name: %s, from: %d to: %d, error: %v",
-					podArgs.PodName, podArgs.PodTrafficMonitorState, PodTerminated, err)
-				continue
-			}
-
-			err = d.StopApiDumpProcess(podUID, errors.Errorf("pod %s has stopped running, status: %s", podArgs.PodName, podStatus))
-			if err != nil {
-				printer.Errorf("Failed to stop api dump process, pod name: %s, error: %v", podArgs.PodName, err)
-			}
-		}
-	}
-}
-
 func (d *Daemonset) PodsHealthWorker(done <-chan struct{}) {
 	if d.PodHealthCheckInterval <= 0 {
 		return
@@ -306,96 +255,4 @@ func (d *Daemonset) PodsHealthWorker(done <-chan struct{}) {
 			d.checkPodsHealth()
 		}
 	}
-
-}
-
-func (d *Daemonset) StartApiDumpProcess(podUID types.UID) error {
-	podArgs, err := d.getPodArgsFromMap(podUID)
-	if err != nil {
-		return err
-	}
-
-	err = podArgs.changePodTrafficMonitorState(TrafficMonitoringStarted, PodDetected, PodInitialized)
-	if err != nil {
-		return errors.Wrapf(err, "failed to change pod state, pod name: %s, from: %d to: %d",
-			podArgs.PodName, podArgs.PodTrafficMonitorState, TrafficMonitoringStopped)
-	}
-
-	go func() (funcErr error) {
-		// defer function handle the error (if any) in the apidump process and change the pod state accordingly
-		defer func() {
-			nextState := TrafficMonitoringEnded
-
-			if err := recover(); err != nil {
-				printer.Errorf("Panic occurred in apidump process for pod %s, err: %v\n%v\n",
-					podArgs.PodName, err, string(debug.Stack()))
-				nextState = TrafficMonitoringFailed
-			} else if funcErr != nil {
-				printer.Errorf("Error occurred in apidump process for pod %s, err: %v", podArgs.PodName, funcErr)
-				nextState = TrafficMonitoringFailed
-			} else {
-				printer.Infof("Apidump process ended for pod %s", podArgs.PodName)
-			}
-
-			err = podArgs.changePodTrafficMonitorState(nextState, TrafficMonitoringStarted)
-			if err != nil {
-				printer.Errorf("Failed to change pod state, pod name: %s, from: %d to: %d, error: %v",
-					podArgs.PodName, podArgs.PodTrafficMonitorState, nextState, err)
-				return
-			}
-
-			// It is possible that the apidump process is already stopped and the stopChannel is of no use
-			// This is just a safety check
-			err := d.StopApiDumpProcess(podUID, err)
-			if err != nil {
-				printer.Errorf("Failed to stop api dump process, pod name: %s, error: %v", podArgs.PodName, err)
-			}
-		}()
-
-		networkNamespace, err := d.CRIClient.GetNetworkNamespace(podArgs.ContainerUUID)
-		if err != nil {
-			funcErr = errors.Errorf("Failed to get network namespace for pod/containerUUID: %s/%s, err: %v",
-				podArgs.PodName, podArgs.ContainerUUID, err)
-			return
-		}
-
-		apidumpArgs := apidump.Args{
-			ClientID:  telemetry.GetClientID(),
-			Domain:    rest.Domain,
-			ServiceID: podArgs.InsightsProjectID,
-			ReproMode: d.InsightsReproModeEnabled,
-			DaemonsetArgs: optionals.Some(apidump.DaemonsetArgs{
-				TargetNetworkNamespaceOpt: networkNamespace,
-				StopChan:                  podArgs.StopChan,
-				APIKey:                    podArgs.PodCreds.InsightsAPIKey,
-				Environment:               podArgs.PodCreds.InsightsEnvironment,
-			}),
-		}
-
-		if err := apidump.Run(apidumpArgs); err != nil {
-			funcErr = errors.Wrapf(err, "Failed to run apidump process for pod %s", podArgs.PodName)
-		}
-		return
-	}()
-
-	return nil
-}
-
-func (d *Daemonset) StopApiDumpProcess(podUID types.UID, err error) error {
-	podArgs, err := d.getPodArgsFromMap(podUID)
-	if err != nil {
-		return err
-	}
-
-	err = podArgs.changePodTrafficMonitorState(TrafficMonitoringStopped,
-		PodTerminated, TrafficMonitoringFailed, TrafficMonitoringEnded)
-	if err != nil {
-		return errors.Wrapf(err, "failed to change pod state, pod name: %s, from: %d to: %d",
-			podArgs.PodName, podArgs.PodTrafficMonitorState, TrafficMonitoringStopped)
-	}
-
-	printer.Infof("Stopping API dump process for pod %s", podArgs.PodName)
-	podArgs.StopChan <- err
-
-	return nil
 }
